requirements.txt
langchain
langchain-groq
faiss-cpu
sentence-transformers
pypdf
python-dotenv
langchain-community

